Changed floating point to double and fixed warnings

+++ HW07/Alarm.h
@@ -9,9 +9,9 @@ public:
     virtual ~Alarm( );
     virtual void notify(double,double);
 private:
-    float x;
-    float y;
-    float soundAlarm;
+    double x;
+    double y;
+    double soundAlarm;
     int id;
 };
